requirements.txt
-e .
beautifulsoup4
datasette
datasette-publish-fly
datasette-render-markdown
html5lib
ipdb
isort
jupyter
lxml
mypy
nb_black
numpy
pandas
pdfplumber
pre-commit
puzpy
pytest
pytest-mock
requests
sqlite-utils
tqdm
